+++ docker/haru/src/app/DDD/Reservation/DateFormatter.php
@@ -1,0 +1,14 @@
+<?php
+// 日付フォーマット
+namespace App\DDD\Reservation;
+
+class DateFormatter
+{
+    // 'Y-m-d'形式に変換
+    public function formatDateTime($dateTime)
+    {
+        return $dateTime instanceof \DateTimeInterface
+            ? $dateTime->format('Y-m-d')
+            : (new \DateTime($dateTime))->format('Y-m-d');
+    }
+}
